Finished PlayerMove and Dash, Jump, and Smoke
Finished the player movement with mouse movement and jump. Added the Updraft ability and also added the slow fall when holding space. Added the Dash ability and dash whichever way you are looking and/or moving. Added the smoke, when holding the button you control it and it also disappears after 3 seconds.